MAINT: move DTconfig code to where other DT classes are
