+++ scpdt/_util.py
@@ -9,93 +9,6 @@ import tempfile
 import inspect
 from contextlib import contextmanager
 
-from doctest import NORMALIZE_WHITESPACE, ELLIPSIS, IGNORE_EXCEPTION_DETAIL
-
-import numpy as np
-
-
-class DTConfig:
-    """A bag class to collect various configuration bits. 
-
-    If an attribute is None, helpful defaults are subsituted.
-
-    Attributes
-    ----------
-    default_namespace : dict
-        The namespace to run examples in
-    check_namespace : dict
-        The namespace to do checks in
-    rndm_markers : set
-        Additional directives which act like `# doctest: + SKIP`
-    optionflags : int
-        doctest optionflags
-    stopwords : list
-        If an example contains any of these stopwords, do not check the output
-        (but do check that the source is valid python)
-
-    """
-    def __init__(self, *, default_namespace=None, check_namespace=None,
-                          rndm_markers=None,
-                          # DTRunner configuration
-                          optionflags=None,
-                          # DTFinder configuration
-                          stopwords=None):
-
-        ### DTChecker configuration ###
-
-        # The namespace to run examples in
-        if default_namespace is None:
-            default_namespace = {'np': np}
-        self.default_namespace = default_namespace
-
-        # The namespace to do checks in
-        if check_namespace is None:
-            check_namespace = {
-                  'np': np,
-                  'assert_allclose': np.testing.assert_allclose,
-                  'assert_equal': np.testing.assert_equal,
-                  # recognize numpy repr's
-                  'array': np.array,
-                  'matrix': np.matrix,
-                  'masked_array': np.ma.masked_array,
-                  'int64': np.int64,
-                  'uint64': np.uint64,
-                  'int8': np.int8,
-                  'int32': np.int32,
-                  'float32': np.float32,
-                  'float64': np.float64,
-                  'dtype': np.dtype,
-                  'nan': np.nan,
-                  'NaN': np.nan,
-                  'inf': np.inf,
-                  'Inf': np.inf,}
-            self.check_namespace = check_namespace
-
-        # Additional directives which act like `# doctest: + SKIP`
-        if rndm_markers is None:
-            rndm_markers = {'# random', '# Random',
-                            '#random', '#Random',
-                            "# may vary"}
-        self.rndm_markers = rndm_markers
-
-        ### DTRunner configuration ###
-
-        # doctest optionflags
-        if optionflags is None:
-            optionflags = NORMALIZE_WHITESPACE | ELLIPSIS | IGNORE_EXCEPTION_DETAIL
-        self.optionflags = optionflags
-
-        ### DTFinder configuration ###
-
-        # ignore examples which contain any of these stopwords
-        if stopwords is None:
-            stopwords = {'plt.', '.hist', '.show', '.ylim', '.subplot(',
-                 'set_title', 'imshow', 'plt.show', '.axis(', '.plot(',
-                 '.bar(', '.title', '.ylabel', '.xlabel', 'set_ylim', 'set_xlim',
-                 '# reformatted', '.set_xlabel(', '.set_ylabel(', '.set_zlabel(',
-                 '.set(xlim=', '.set(ylim=', '.set(xlabel=', '.set(ylabel=', '.xlim('}
-        self.stopwords = stopwords
-
 
 @contextmanager
 def matplotlib_make_nongui():
